requirements.txt
transformers
torch
requests
pillow
python-dotenv